Attempts to solve one bug - solves another.
Tried to fix a strange bug where patch has an error somewhere in
between the last .then() in it's promise chain and the .catch()

(I found this out by logging arbitrary messages in each step of the
chain.)

The problem I solved without meaning to was one where show was
erroring out because it was trying to check the value of chain
that hadn't yet been made. I fixed it w/ a ternary statement
that checked whether or not the chains array was populated
before calling the chain's property.

May need to issue queue the first error. Will see if I can crack it
on my own one last time.