Fixed dysfunctional middleware: breakOldChains
Used a different pattern, whereby the response from a document query
is wrapped in a cursor - a stream-compatible interface - so that
each doc is pulled consecutively from the data base and converted
into a Promise, and then the next doc is pulled only when the
preceding is resolved. This allows for conservation of work:
if an error is thrown early, the stream is closed. In addition,
since docs are being called asynchronously, it allows for more
atomic control over the action (in this case, update) that you
are trying to perform.